moved all logic to lib.rs
