+++ Cargo.toml
@@ -25,5 +25,5 @@ memmap2 = "0.9.9"
 [profile.release]
 codegen-units = 1
 lto = true
-# strip = true
-debug = true
+strip = true
+#debug = true
